+++ girder/api/rest.py
@@ -29,6 +29,7 @@ import sys
 import traceback
 import unicodedata
 
+from dogpile.cache.util import kwarg_function_key_generator
 from . import docs
 from girder import events, logger, logprint
 from girder.constants import SettingKey, TokenScope, SortDir
@@ -38,6 +39,7 @@ from girder.models.setting import Setting
 from girder.models.token import Token
 from girder.models.user import User
 from girder.utility import toBool, config, JsonEncoder, optionalArgumentDecorator
+from girder.utility._cache import requestCache
 from girder.utility.model_importer import ModelImporter
 from six.moves import range, urllib
 
@@ -141,45 +143,8 @@ def iterBody(length=READ_BUFFER_LEN, strictLength=False):
             yield buf
 
 
-def _cacheAuthUser(fun):
-    """
-    This decorator for getCurrentUser ensures that the authentication procedure
-    is only performed once per request, and is cached on the request for
-    subsequent calls to getCurrentUser().
-    """
-    def inner(returnToken=False, *args, **kwargs):
-        if not returnToken and hasattr(cherrypy.request, 'girderUser'):
-            return cherrypy.request.girderUser
-
-        user = fun(returnToken, *args, **kwargs)
-        if isinstance(user, tuple):
-            setCurrentUser(user[0])
-        else:
-            setCurrentUser(user)
-
-        return user
-    return inner
-
-
-def _cacheAuthToken(fun):
-    """
-    This decorator for getCurrentToken ensures that the token lookup
-    is only performed once per request, and is cached on the request for
-    subsequent calls to getCurrentToken().
-    """
-    def inner(*args, **kwargs):
-        if hasattr(cherrypy.request, 'girderToken'):
-            return cherrypy.request.girderToken
-
-        token = fun(*args, **kwargs)
-        setattr(cherrypy.request, 'girderToken', token)
-
-        return token
-    return inner
-
-
-@_cacheAuthToken
-def getCurrentToken(allowCookie=False):
+@requestCache.cache_on_arguments(function_key_generator=kwarg_function_key_generator)
+def getCurrentToken(allowCookie=None):
     """
     Returns the current valid token object that was passed via the token header
     or parameter, or None if no valid token was passed.
@@ -190,9 +155,13 @@ def getCurrentToken(allowCookie=False):
         This should only be used on read-only operations that will not make any
         changes to data on the server, and only in cases where the user agent
         behavior makes passing custom headers infeasible, such as downloading
-        data to disk in the browser.
+        data to disk in the browser. In the event that allowCookie is not explicitly
+        passed, it will default to False unless the access.cookie decorator is used.
     :type allowCookie: bool
     """
+    if allowCookie is None:
+        allowCookie = getattr(cherrypy.request, 'girderAllowCookie', False)
+
     tokenStr = None
     if 'token' in cherrypy.request.params:  # Token as a parameter
         tokenStr = cherrypy.request.params.get('token')
@@ -207,7 +176,6 @@ def getCurrentToken(allowCookie=False):
     return Token().load(tokenStr, force=True, objectId=False)
 
 
-@_cacheAuthUser
 def getCurrentUser(returnToken=False):
     """
     Returns the currently authenticated user based on the token header or
@@ -220,6 +188,9 @@ def getCurrentUser(returnToken=False):
               logged in or the token is invalid or expired.  If
               returnToken=True, returns a tuple of (user, token).
     """
+    if not returnToken and hasattr(cherrypy.request, 'girderUser'):
+        return cherrypy.request.girderUser
+
     event = events.trigger('auth.user.get')
     if event.defaultPrevented and len(event.responses) > 0:
         return event.responses[0]
@@ -227,6 +198,8 @@ def getCurrentUser(returnToken=False):
     token = getCurrentToken()
 
     def retVal(user, token):
+        setCurrentUser(user)
+
         if returnToken:
             return user, token
         else:
@@ -920,10 +893,8 @@ class Resource(ModelImporter):
                 forceCookie = False
             if cookieAuth:
                 if forceCookie or method in ('head', 'get'):
-                    # getCurrentToken will cache its output, so calling it
-                    # once with allowCookie will make the parameter
-                    # effectively permanent (for the request)
-                    getCurrentToken(allowCookie=True)
+                    # Allow cookies for the rest of the request
+                    setattr(cherrypy.request, 'girderAllowCookie', True)
 
         kwargs['params'] = params
         # Add before call for the API method. Listeners can return